backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.10
python-dotenv==1.0.1
kafka-python-ng
redis==5.0.7
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.8.0
alembic==1.13.2
pytest==8.3.2